session repository fix
