+++ internal/api/auth-handler.go
@@ -2,7 +2,6 @@ package api
 
 import (
 	"authorization-service/pkg/models"
-	"fmt"
 
 	"github.com/labstack/echo"
 )
@@ -41,15 +40,4 @@ func (api *API) Authenticate(c echo.Context) error {
 		}
 	}
 	return c.JSON(500, models.ErrorResponse{Code: 004, Message: "user not logged due api error"})
-}
-
-// CheckValidSession : we check if the session is valid
-func (api *API) CheckValidSession(c echo.Context) error {
-	cookie, err := c.Cookie("HelpSasterAuth")
-	if err == nil {
-		c.JSON(500, models.ErrorResponse{Code: 005, Message: "cookie not found need to relogin"})
-	}
-	fmt.Println(cookie.Name)
-	fmt.Println(cookie.Value)
-	return c.JSON(500, models.ErrorResponse{Code: 006, Message: "session could not be checked due api error"})
 }
